refactor(truecloud): consolidate ad removal patches and update package
Consolidates multiple individual ad-related patches (ads, cloud boot, cloud API, help robot) into a single `TrueCloudAdRemovalPatch`.
Moves patches from `app.template.patches.truecloud` to `app.truecloud.patches`.
Note: The `RemoveEventCloudPopupPatch` functionality for event tab popups and banners is no longer included in the consolidated patch.
