+++ patches/src/main/kotlin/app/truecloud/patches/shared/Constants.kt
@@ -1,4 +1,4 @@
-package app.template.patches.shared
+package app.truecloud.patches.shared
 
 import app.morphe.patcher.patch.ApkFileType
 import app.morphe.patcher.patch.AppTarget
